feat(efile): implement phase 1 foundation
