+++ .env.example
@@ -1,0 +1,2 @@
+VITE_EFILE_BASE_URL=https://api.uslegalpro.com/v4
+VITE_EFILE_CLIENT_TOKEN=YOUR_CLIENT_TOKEN
