+++ .gitignore
@@ -1,1 +1,4 @@
 **/.claude/settings.local.json
+.env.local
+node_modules/
+dist/
