+++ docs/api/e-filing/OFFICIAL-API-DOCUMENTATION-WITH-CREDENTIALS.txt
@@ -1,8 +1,8 @@
 ====================================================================================================================================================
 #1 Credentials:
-Client token is: EVICT87
-Stage username: your_tyler_email@example.com
-Stage password: your_tyler_password
+Client token is: <CLIENT_TOKEN>
+Stage username: <USERNAME>
+Stage password: <PASSWORD>
 ====================================================================================================================================================
 #2 How to eFile:
