+++ package.json
@@ -35,7 +35,8 @@
     "tailwind-merge": "^3.2.0",
     "uuid": "^9.0.1",
     "xlsx": "^0.18.5",
-    "zod": "^3.22.4"
+    "zod": "^3.22.4",
+    "axios": "^1.6.0"
   },
   "devDependencies": {
     "@babel/parser": "^7.23.6",
