+++ src/components/efile/EFileUploadItem.tsx
@@ -1,0 +1,3 @@
+// Placeholder for future file upload item component
+const EFileUploadItem = () => null;
+export default EFileUploadItem;
