+++ src/utils/efile/auth.ts
@@ -1,0 +1,31 @@
+import type { AuthenticateRequest, AuthenticateResponse } from '@/types/efile';
+import { apiClient } from './apiClient';
+
+const CLIENT_TOKEN = import.meta.env.VITE_EFILE_CLIENT_TOKEN;
+
+export async function authenticate(
+  username: string,
+  password: string,
+): Promise<string> {
+  const req: AuthenticateRequest = { data: { username, password } };
+  const { data } = await apiClient.post<AuthenticateResponse>(
+    '/il/user/authenticate',
+    req,
+    { headers: { clienttoken: CLIENT_TOKEN } },
+  );
+  return data.item.auth_token;
+}
+
+export function storeToken(token: string, expiresIn: number) {
+  const expires = Date.now() + expiresIn * 1000;
+  localStorage.setItem('efileAuth', JSON.stringify({ token, expires }));
+}
+
+export function getStoredToken(): { token: string; expires: number } | null {
+  const raw = localStorage.getItem('efileAuth');
+  return raw ? JSON.parse(raw) : null;
+}
+
+export function isTokenExpired(expires: number) {
+  return Date.now() > expires;
+}
